requirements.txt
gspread==5.7.1
mysql_connector_repackaged==0.3.1
pandas==1.5.1
plotly==5.11.0
streamlit==1.15.0
